src/mini_solver_backend/services/ai_processing_engine/requirements.txt
google-genai
aio-pika
kink
pydantic
python-dotenv
google-cloud-storage
google-cloud-firestore
google-auth
google-auth-oauthlib